+++ dialects/postgresql/postgresql.go
@@ -9,6 +9,10 @@ import (
 
 type postgresql struct{}
 
+func (postgresql) String() string {
+	return "postgresql"
+}
+
 func (postgresql) Placeholder(index int) string {
 	return "$" + strconv.Itoa(index)
 }
@@ -27,6 +31,14 @@ func (postgresql) QuoteIdentifier(identifier string) string {
 
 func (postgresql) LastInsertIdMethod() reform.LastInsertIdMethod {
 	return reform.Returning
+}
+
+func (postgresql) SelectLimitMethod() reform.SelectLimitMethod {
+	return reform.Limit
+}
+
+func (postgresql) DefaultValuesMethod() reform.DefaultValuesMethod {
+	return reform.DefaultValues
 }
 
 // Dialect implements reform.Dialect for PostgreSQL.