+++ querier_commands.go
@@ -5,22 +5,28 @@ import (
 	"strings"
 )
 
-func filteredColumnsAndValues(record Record, columnsIn []string, isUpdate bool) (columns []string, values []interface{}, err error) {
+func filteredColumnsAndValues(str Struct, columnsIn []string, isUpdate bool) (columns []string, values []interface{}, err error) {
 	columnsSet := make(map[string]struct{}, len(columnsIn))
 	for _, c := range columnsIn {
 		columnsSet[c] = struct{}{}
 	}
 
 	// select columns from set and collect values
-	table := record.Table()
-	pk := int(table.PKColumnIndex())
-	allColumns := table.Columns()
-	allValues := record.Values()
+	view := str.View()
+	allColumns := view.Columns()
+	allValues := str.Values()
 	columns = make([]string, 0, len(columnsSet))
 	values = make([]interface{}, 0, len(columns))
+
+	record, _ := str.(Record)
+	var pk uint
+	if record != nil {
+		pk = view.(Table).PKColumnIndex()
+	}
+
 	for i, c := range allColumns {
 		if _, ok := columnsSet[c]; ok {
-			if isUpdate && i == pk {
+			if isUpdate && record != nil && i == int(pk) {
 				err = fmt.Errorf("reform: will not update PK column: %s", c)
 				return
 			}
@@ -52,20 +58,38 @@ func (q *Querier) insert(str Struct, columns []string, values []interface{}) err
 
 	view := str.View()
 	record, _ := str.(Record)
+	lastInsertIdMethod := q.LastInsertIdMethod()
+	defaultValuesMethod := q.DefaultValuesMethod()
 
-	query := fmt.Sprintf("INSERT INTO %s (%s) VALUES (%s)",
-		q.QualifiedView(view),
-		strings.Join(columns, ", "),
-		strings.Join(placeholders, ", "),
-	)
+	var pk uint
+	if record != nil {
+		pk = view.(Table).PKColumnIndex()
+	}
 
-	switch q.Dialect.LastInsertIdMethod() {
+	// make query
+	query := "INSERT INTO " + q.QualifiedView(view)
+	if len(columns) != 0 || defaultValuesMethod == EmptyLists {
+		query += " (" + strings.Join(columns, ", ") + ")"
+	}
+	if record != nil && lastInsertIdMethod == OutputInserted {
+		query += fmt.Sprintf(" OUTPUT INSERTED.%s", q.QuoteIdentifier(view.Columns()[pk]))
+	}
+	if len(placeholders) != 0 || defaultValuesMethod == EmptyLists {
+		query += fmt.Sprintf(" VALUES (%s)", strings.Join(placeholders, ", "))
+	} else {
+		query += " DEFAULT VALUES"
+	}
+	if record != nil && lastInsertIdMethod == Returning {
+		query += fmt.Sprintf(" RETURNING %s", q.QuoteIdentifier(view.Columns()[pk]))
+	}
+
+	switch lastInsertIdMethod {
 	case LastInsertId:
 		res, err := q.Exec(query, values...)
 		if err != nil {
 			return err
 		}
-		if record != nil {
+		if record != nil && !record.HasPK() {
 			id, err := res.LastInsertId()
 			if err != nil {
 				return err
@@ -74,12 +98,9 @@ func (q *Querier) insert(str Struct, columns []string, values []interface{}) err
 		}
 		return nil
 
-	case Returning:
+	case Returning, OutputInserted:
 		var err error
 		if record != nil {
-			pk := view.(Table).PKColumnIndex()
-
-			query += fmt.Sprintf(" RETURNING %s", q.QuoteIdentifier(view.Columns()[pk]))
 			err = q.QueryRow(query, values...).Scan(record.PKPointer())
 		} else {
 			_, err = q.Exec(query, values...)
@@ -116,10 +137,9 @@ func (q *Querier) Insert(str Struct) error {
 	values := str.Values()
 	columns := view.Columns()
 	record, _ := str.(Record)
-	var pk uint
 
 	if record != nil {
-		pk = view.(Table).PKColumnIndex()
+		pk := view.(Table).PKColumnIndex()
 
 		// cut primary key
 		if !record.HasPK() {
@@ -137,19 +157,17 @@ func (q *Querier) Insert(str Struct) error {
 //
 // It fills record's primary key field.
 func (q *Querier) InsertColumns(str Struct, columns ...string) error {
-	record, _ := str.(Record)
-
-	err := q.beforeInsert(record)
+	err := q.beforeInsert(str)
 	if err != nil {
 		return err
 	}
 
-	columns, values, err := filteredColumnsAndValues(record, columns, false)
+	columns, values, err := filteredColumnsAndValues(str, columns, false)
 	if err != nil {
 		return err
 	}
 
-	return q.insert(record, columns, values)
+	return q.insert(str, columns, values)
 }
 
 // InsertMulti inserts several structs into SQL database table with single query.
@@ -262,7 +280,7 @@ func (q *Querier) update(record Record, columns []string, values []interface{}) 
 		return ErrNoRows
 	}
 	if ra > 1 {
-		panic(fmt.Errorf("reform: %d rows by UPDATE by primary key. Please report this bug.", ra))
+		panic(fmt.Sprintf("reform: %d rows by UPDATE by primary key. Please report this bug.", ra))
 	}
 	return nil
 }
@@ -373,7 +391,7 @@ func (q *Querier) Delete(record Record) error {
 		return ErrNoRows
 	}
 	if ra > 1 {
-		panic(fmt.Errorf("reform: %d rows by DELETE by primary key. Please report this bug.", ra))
+		panic(fmt.Sprintf("reform: %d rows by DELETE by primary key. Please report this bug.", ra))
 	}
 	return nil
 }